+++ fixtures/golden/case/05-nested-function-calls-in-when/expected.sql
@@ -15,10 +15,8 @@ select
                 substr(date_add(confirm_time, 729), 1, 10),
                 cast(last_day(concat(report_month, '-01')) as string)
             ),
-            greatest(
-                substr(confirm_time, 1, 10),
-                concat(report_month, '-01')
-            )
-        ) + 1
+            greatest(substr(confirm_time, 1, 10), concat(report_month, '-01'))
+        )
+        + 1
     end as consumed_days
 from style_lab.complex_case_source;